refactor: 존재하는 task에 대한 회고만 조회하도록 수정
