--- src/main/kotlin/com/ssak3/timeattack/retrospection/repository/entity/RetrospectionRepository.kt
@@ -1,7 +1,0 @@
-package com.ssak3.timeattack.retrospection.repository.entity
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface RetrospectionRepository : JpaRepository<RetrospectionEntity, Long> {
-    fun findAllByMemberId(memberId: Long): List<RetrospectionEntity>
-}
